I2C character display fixes + sample cleanup

+++ Source/Meadow.Foundation.Peripherals/Displays.Lcd.CharacterDisplay/Driver/Displays.Lcd.CharacterDisplay/GpioCharacterDisplay.cs
@@ -206,7 +206,7 @@ namespace Meadow.Foundation.Displays.Lcd
         {
             string screentText = text;
 
-            if (screentText.Length + (int)cursorColumn > DisplayConfig.Width)
+            if (screentText.Length + cursorColumn > DisplayConfig.Width)
             {
                 screentText = screentText.Substring(0, DisplayConfig.Width - cursorColumn);
             }
